refactor: reorganize tracker factory mod
