+++ src/servers/http/v1/services/announce.rs
@@ -32,7 +32,7 @@ mod tests {
 
     use crate::shared::bit_torrent::info_hash::InfoHash;
     use crate::shared::clock::DurationSinceUnixEpoch;
-    use crate::tracker::services::common::tracker_factory;
+    use crate::tracker::services::tracker_factory;
     use crate::tracker::{peer, Tracker};
 
     fn public_tracker() -> Tracker {
