+++ src/servers/http/v1/services/scrape.rs
@@ -41,7 +41,7 @@ mod tests {
 
     use crate::shared::bit_torrent::info_hash::InfoHash;
     use crate::shared::clock::DurationSinceUnixEpoch;
-    use crate::tracker::services::common::tracker_factory;
+    use crate::tracker::services::tracker_factory;
     use crate::tracker::{peer, Tracker};
 
     fn public_tracker() -> Tracker {
