+++ src/tracker/services/statistics.rs
@@ -40,8 +40,8 @@ mod tests {
     use torrust_tracker_test_helpers::configuration;
 
     use crate::tracker;
-    use crate::tracker::services::common::tracker_factory;
     use crate::tracker::services::statistics::{get_metrics, TrackerMetrics};
+    use crate::tracker::services::tracker_factory;
 
     pub fn tracker_configuration() -> Arc<Configuration> {
         Arc::new(configuration::ephemeral())
