+++ src/tracker/services/torrent.rs
@@ -141,9 +141,9 @@ mod tests {
         use torrust_tracker_test_helpers::configuration;
 
         use crate::shared::bit_torrent::info_hash::InfoHash;
-        use crate::tracker::services::common::tracker_factory;
         use crate::tracker::services::torrent::tests::sample_peer;
         use crate::tracker::services::torrent::{get_torrent_info, Info};
+        use crate::tracker::services::tracker_factory;
 
         pub fn tracker_configuration() -> Arc<Configuration> {
             Arc::new(configuration::ephemeral())
@@ -196,9 +196,9 @@ mod tests {
         use torrust_tracker_test_helpers::configuration;
 
         use crate::shared::bit_torrent::info_hash::InfoHash;
-        use crate::tracker::services::common::tracker_factory;
         use crate::tracker::services::torrent::tests::sample_peer;
         use crate::tracker::services::torrent::{get_torrents, BasicInfo, Pagination};
+        use crate::tracker::services::tracker_factory;
 
         pub fn tracker_configuration() -> Arc<Configuration> {
             Arc::new(configuration::ephemeral())
